# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Cytisus API
  version: 0.6.0
  description: Shared client contract for deterministic Paper Securities, simulated banking, custodial simulated Crypto, and the simulated Card MVP.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: https://api.{environment}.cytisus.invalid
    description: Environment placeholder; a real production domain requires human approval.
    variables:
      environment:
        default: local
        enum: [local, test, staging, production]
...
paths:
...
  /v1/card:
    get:
      operationId: getCardProfile
      summary: Return the Card profile and virtual or physical cards
      tags: [Card]
      responses:
        "200":
          description: Card profile with truthful Wallet availability states.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/card/cards:
    post:
      operationId: createCard
      summary: Create or replay a virtual, plastic, or Metal card request
      tags: [Card]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardCreateRequest"
      responses:
        "200":
          description: Existing card replayed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "201":
          description: Card request created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/card/cards/{card_id}/actions:
    post:
      operationId: actOnCard
      summary: Activate, set PIN state, freeze, replace, report, or close a card
      tags: [Card]
      parameters:
        - $ref: "#/components/parameters/CardID"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardActionRequest"
      responses:
        "200":
          description: Card state updated or replayed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/card/spending-power:
    get:
      operationId: getCardSpendingPower
      summary: Return Ledger-backed cash, holds, receivable, and risk-adjusted eligible collateral
      tags: [Card]
      responses:
        "200":
          description: Current Card spending power; formulas are summarized without exposing manipulable thresholds.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardSpendingPower"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/card/repayment-mode:
    post:
      operationId: configureCardRepaymentMode
      summary: Configure CASH_ONLY, CASH_THEN_AUTO_SELL, or MONTHLY_STATEMENT
      tags: [Card]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardRepaymentModeRequest"
      responses:
        "200":
          description: Repayment mode updated or replayed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardProfile"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/card/auto-sell-mandate:
    get:
      operationId: getCardAutoSellMandate
      summary: Return the active pre-authorized Auto-Sell waterfall
      tags: [Card]
      responses:
        "200":
          description: Active mandate.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardAutoSellMandate"
        "409":
          $ref: "#/components/responses/Conflict"
    put:
      operationId: configureCardAutoSellMandate
      summary: Configure assets, order, protected-sale cap, retention, fractionality, and validity
      tags: [Card]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardAutoSellMandateRequest"
      responses:
        "200":
          description: Mandate configured or replayed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardAutoSellMandate"
        "400":
          $ref: "#/components/responses/InvalidRequest"
  /v1/card/authorizations:
    get:
      operationId: listCardAuthorizations
      summary: List NFC, e-commerce, manual, and delayed offline authorizations
      tags: [Card]
      parameters:
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Authorization history.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardAuthorizationList"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/card/captures:
    get:
      operationId: listCardCaptures
      summary: List partial or final captures, FX, tips, repayment, and protected Auto-Sell executions
      tags: [Card]
      parameters:
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Capture history.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardCaptureList"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/card/disputes:
    get:
      operationId: listCardDisputes
      summary: List the authenticated customer's Card disputes
      tags: [Card]
      parameters:
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Dispute history.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardDisputeList"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: openCardDispute
      summary: Open an idempotent Card dispute and compliance case
      tags: [Card]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardDisputeOpenRequest"
      responses:
        "201":
          description: Dispute opened.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardDispute"
        "422":
          $ref: "#/components/responses/Unprocessable"
  /v1/card/statements:
    get:
      operationId: listCardStatements
      summary: List Card monthly statements
      tags: [Card]
      responses:
        "200":
          description: Statement history.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardStatementList"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/card/statements/{statement_id}/pay:
    post:
      operationId: payCardStatement
      summary: Pay an open statement with settled and withdrawable USD through Ledger
      tags: [Card]
      parameters:
        - $ref: "#/components/parameters/StatementID"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Statement paid or replayed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardStatement"
        "422":
          $ref: "#/components/responses/Unprocessable"
  /v1/card/notifications:
    get:
      operationId: listCardNotifications
      summary: List in-app Card notifications emitted with domain mutations
      tags: [Card]
      parameters:
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Notification inbox.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardNotificationList"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /internal/v1/simulators/card/capabilities:
    get:
      operationId: getCardSimulatorCapabilities
      summary: Return truthful local Card and Wallet capability states
      security: []
      tags: [Simulators]
      responses:
        "200":
          description: Local simulator capabilities.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardCapabilities"
        "403":
          $ref: "#/components/responses/Forbidden"
  /internal/v1/simulators/card/collateral:
    post:
      operationId: seedCardCollateral
      summary: Seed a Ledger-backed simulated live securities collateral fixture
      tags: [Simulators]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardCollateralSeedRequest"
      responses:
        "201":
          description: Collateral posted through Ledger.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          $ref: "#/components/responses/InvalidRequest"
  /internal/v1/simulators/card/quotes:
    post:
      operationId: updateCardCollateralQuote
      summary: Set a SIMULATED, STALE, or unavailable Card reference-price fixture
      security: []
      tags: [Simulators]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardQuoteUpdateRequest"
      responses:
        "200":
          description: Quote fixture updated.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          $ref: "#/components/responses/InvalidRequest"
  /internal/v1/simulators/card/terminal/authorizations:
    post:
      operationId: simulateCardAuthorization
      summary: Run the local merchant terminal or NFC authorization simulator
      tags: [Simulators]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardAuthorizationRequest"
      responses:
        "200":
          description: Approved or declined authorization; duplicate identifiers replay deterministically.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardAuthorization"
        "409":
          $ref: "#/components/responses/Conflict"
  /internal/v1/simulators/card/terminal/captures:
    post:
      operationId: simulateCardCapture
      summary: Capture all or part of an authorization with clearing FX and optional tip
      tags: [Simulators]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardCaptureRequest"
      responses:
        "200":
          description: Capture posted through Ledger and repaid according to policy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardCapture"
        "422":
          $ref: "#/components/responses/Unprocessable"
  /internal/v1/simulators/card/terminal/reversals:
    post:
      operationId: simulateCardReversal
      summary: Release an authorization hold through Ledger
      tags: [Simulators]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardReversalRequest"
      responses:
        "200":
          description: Hold reversed or duplicate replayed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardAuthorization"
        "400":
          $ref: "#/components/responses/InvalidRequest"
  /internal/v1/simulators/card/terminal/refunds:
    post:
      operationId: simulateCardRefund
      summary: Apply a refund against receivable first, then eligible cash
      tags: [Simulators]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardRefundRequest"
      responses:
        "200":
          description: Refund posted or duplicate replayed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardRefund"
        "400":
          $ref: "#/components/responses/InvalidRequest"
  /internal/v1/admin/card/disputes:
    get:
      operationId: listAdminCardDisputes
      summary: List Card disputes for authorized review roles
      security:
        - adminID: []
          adminRole: []
      tags: [Card admin]
      parameters:
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Card dispute review queue.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardDisputeList"
        "403":
          $ref: "#/components/responses/Forbidden"
  /internal/v1/admin/card/disputes/{dispute_id}/resolve:
    post:
      operationId: resolveAdminCardDispute
      summary: Accept or reject a dispute with audit, compliance transition, notification, and optional Ledger credit
      security:
        - adminID: []
          adminRole: []
      tags: [Card admin]
      parameters:
        - $ref: "#/components/parameters/DisputeID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardDisputeResolveRequest"
      responses:
        "200":
          description: Dispute resolved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardDispute"
        "403":
          $ref: "#/components/responses/Forbidden"
  /internal/v1/admin/card/cards/{card_id}/lifecycle:
    post:
      operationId: advancePhysicalCardLifecycle
      summary: Advance plastic or Metal manufacturing and delivery state
      security:
        - adminID: []
          adminRole: []
      tags: [Card admin]
      parameters:
        - $ref: "#/components/parameters/CardID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardPhysicalLifecycleRequest"
      responses:
        "200":
          description: Physical card lifecycle advanced.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "403":
          $ref: "#/components/responses/Forbidden"
  /internal/v1/admin/card/statements:
    post:
      operationId: generateCardStatement
      summary: Generate an open MONTHLY_STATEMENT projection from Card receivable
      security:
        - adminID: []
          adminRole: []
      tags: [Card admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardStatementGenerateRequest"
      responses:
        "201":
          description: Statement generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardStatement"
        "403":
          $ref: "#/components/responses/Forbidden"
  /internal/v1/admin/card/reconciliation:
    get:
      operationId: listCardReconciliationRuns
      summary: List Card provider-to-Ledger reconciliation runs
      security:
        - adminID: []
          adminRole: []
      tags: [Card admin]
      parameters:
        - name: customer_reference
          in: query
          required: false
          schema:
            type: string
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Reconciliation history.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardReconciliationList"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: reconcileCardProvider
      summary: Reconcile provider holds and receivables without automatic balance adjustment
      security:
        - adminID: []
          adminRole: []
      tags: [Card admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardReconciliationRequest"
      responses:
        "200":
          description: Reconciliation result; a difference creates a visible case.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardReconciliation"
        "403":
          $ref: "#/components/responses/Forbidden"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: OpaquePaperSession
    adminID:
      type: apiKey
      in: header
      name: X-Admin-ID
      description: Synthetic local admin identity. Production requires isolated enterprise SSO.
    adminRole:
      type: apiKey
      in: header
      name: X-Admin-Role
      description: Synthetic local role used for permission tests.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Stable key reused only for retries of the same financial command.
      schema:
        type: string
        minLength: 8
        maxLength: 256
        pattern: "^[a-zA-Z0-9][a-zA-Z0-9._:-]{7,255}$"
...
    CardID:
      name: card_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    InvalidRequest:
      description: The request did not pass schema or domain validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: A valid paper session bearer token is required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Conflict:
      description: Idempotency, replay, quote, or order state conflict.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    CardCreateRequest:
      type: object
      additionalProperties: false
      required: [card_type]
      properties:
        card_type:
          type: string
          enum: [VIRTUAL, PLASTIC, METAL]
    CardActionRequest:
      type: object
      additionalProperties: false
      required: [action]
      properties:
        action:
          type: string
          enum:
            [
              ACTIVATE,
              SET_PIN,
              CHANGE_PIN,
              FREEZE,
              UNFREEZE,
              REPLACE,
              REPORT_LOST,
              REPORT_STOLEN,
              CLOSE,
            ]
        reason_code:
          type: string
    CardRepaymentModeRequest:
      type: object
      additionalProperties: false
      required: [repayment_mode]
      properties:
        repayment_mode:
          type: string
          enum: [CASH_ONLY, CASH_THEN_AUTO_SELL, MONTHLY_STATEMENT]
...
    Card:
      type: object
      additionalProperties: false
      required:
        [
          id,
          card_type,
          status,
          display_name,
          last4,
          pin_set,
          provider,
          provider_card_reference,
          apple_wallet_status,
          google_wallet_status,
          replacement_for_card_id,
          policy_version,
          replayed,
          created_at,
          updated_at,
        ]
      properties:
        id:
          type: string
          format: uuid
        card_type:
          type: string
          enum: [VIRTUAL, PLASTIC, METAL]
        status:
          type: string
        display_name:
          type: string
        last4:
          type: string
          pattern: "^[0-9]{4}$"
        pin_set:
          type: boolean
        provider:
          type: string
          const: local-card-simulator
        provider_card_reference:
          type: string
        apple_wallet_status:
          type: string
          const: UNAVAILABLE_SIMULATOR
        google_wallet_status:
          type: string
          const: UNAVAILABLE_SIMULATOR
        replacement_for_card_id:
          type: [string, "null"]
          format: uuid
        policy_version:
          type: string
        replayed:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CardProfile:
      type: object
      additionalProperties: false
      required:
        [
          customer_reference,
          repayment_mode,
          spending_status,
          spending_status_reason,
          policy_version,
          cards,
          mode,
        ]
      properties:
        customer_reference:
          type: string
        repayment_mode:
          type: string
          enum: [CASH_ONLY, CASH_THEN_AUTO_SELL, MONTHLY_STATEMENT]
        spending_status:
          type: string
          enum: [ACTIVE, FROZEN]
        spending_status_reason:
          type: [string, "null"]
        policy_version:
          type: string
        cards:
          type: array
          items:
            $ref: "#/components/schemas/Card"
        mode:
          type: string
          const: SIMULATED
...
    CardAutoSellMandate:
      type: object
      additionalProperties: false
      required:
        [
          id,
          enabled,
          allow_fractional,
          daily_max_usd,
          valid_until,
          policy_version,
          assets,
          updated_at,
        ]
      properties:
        id:
          type: string
          format: uuid
        enabled:
          type: boolean
        allow_fractional:
          type: boolean
        daily_max_usd:
          $ref: "#/components/schemas/Decimal"
        valid_until:
          type: string
          format: date-time
        policy_version:
          type: string
        assets:
          type: array
          items:
            $ref: "#/components/schemas/CardAutoSellMandateAsset"
        updated_at:
          type: string
          format: date-time
...
security:
  - bearerAuth: []